Assign7 (#63)
* pushed changes

* Update README.md

* Feature (#3)

* commited packer workflow

* added workflow files

* modified packer file

* committed amibuild workflow

* added build ami workflow

* Feature (#1)

* pushed changes

* Update README.md

* commited packer workflow

* added workflow files

* modified packer file

* committed amibuild workflow

* added build ami workflow

* added build ami workflow

* added files

* added

* modified packer file

* packer file modified

* Assignment5 (#4)

* commited packer workflow

* added workflow files

* modified packer file

* committed amibuild workflow

* added build ami workflow

* Feature (#1)

* pushed changes

* Update README.md

* commited packer workflow

* added workflow files

* modified packer file

* committed amibuild workflow

* added build ami workflow

* added build ami workflow

* Feature (#2)

* pushed changes

* Update README.md

* commited packer workflow

* added workflow files

* modified packer file

* committed amibuild workflow

* added build ami workflow

* added build ami workflow

* added files

* added

* Feature (#3)

* pushed changes

* Update README.md

* commited packer workflow

* added workflow files

* modified packer file

* committed amibuild workflow

* added build ami workflow

* added build ami workflow

* added files

* added

* Update ami.us-east-1.pkr.hcl (#4)

* modified packer file

* packer file modified

* committed packer file

* Feature (#6)

* pushed changes

* Update README.md

* commited packer workflow

* added workflow files

* modified packer file

* committed amibuild workflow

* added build ami workflow

* added build ami workflow

* added files

* added

* modified packer file

* packer file modified

* committed packer file

* committed packer file

* Feature (#7)

* pushed changes

* Update README.md

* commited packer workflow

* added workflow files

* modified packer file

* committed amibuild workflow

* added build ami workflow

* added build ami workflow

* added files

* added

* modified packer file

* packer file modified

* committed packer file

* committed packer file

* Update build-ami.yml

* workflows edited

* Feature (#5)

* commited packer workflow

* added workflow files

* modified packer file

* committed amibuild workflow

* added build ami workflow

* Feature (#1)

* pushed changes

* Update README.md

* commited packer workflow

* added workflow files

* modified packer file

* committed amibuild workflow

* added build ami workflow

* added build ami workflow

* Feature (#2)

* pushed changes

* Update README.md

* commited packer workflow

* added workflow files

* modified packer file

* committed amibuild workflow

* added build ami workflow

* added build ami workflow

* added files

* added

* Feature (#3)

* pushed changes

* Update README.md

* commited packer workflow

* added workflow files

* modified packer file

* committed amibuild workflow

* added build ami workflow

* added build ami workflow

* added files

* added

* Update ami.us-east-1.pkr.hcl (#4)

* modified packer file

* packer file modified

* committed packer file

* Feature (#6)

* pushed changes

* Update README.md

* commited packer workflow

* added workflow files

* modified packer file

* committed amibuild workflow

* added build ami workflow

* added build ami workflow

* added files

* added

* modified packer file

* packer file modified

* committed packer file

* committed packer file

* Feature (#7)

* pushed changes

* Update README.md

* commited packer workflow

* added workflow files

* modified packer file

* committed amibuild workflow

* added build ami workflow

* added build ami workflow

* added files

* added

* modified packer file

* packer file modified

* committed packer file

* committed packer file

* Update build-ami.yml

* workflows edited

* Feature (#9)

* pushed changes

* Update README.md

* commited packer workflow

* added workflow files

* modified packer file

* committed amibuild workflow

* added build ami workflow

* added build ami workflow

* added files

* added

* modified packer file

* packer file modified

* committed packer file

* committed packer file

* workflows edited

* Feature (#11)

* pushed changes

* Update README.md

* commited packer workflow

* added workflow files

* modified packer file

* committed amibuild workflow

* added build ami workflow

* added build ami workflow

* added files

* added

* modified packer file

* packer file modified

* committed packer file

* committed packer file

* workflows edited

* added dependencies

* Update setup.sh (#6)

* Update setup.sh (#7)

* added setup files

* setup files added cloudwatch

* Update setup.sh

* Update setup.sh

* updated sh files

* added cloudwatch config

* add ssetup files

* changes in bash file

* small change

* added setup files

* added setup.sh

* added setup

* webapplication changes

* added log file path

* added statsd configurations

* added cjs file

* cjs files

* Update setup.sh

* added file fir json

* add

* added

* config

* added

+++ setup.sh
@@ -12,9 +12,20 @@ sudo DEBIAN_FRONTEND=noninteractive apt install -y unzip
 npm install winston
 
 sudo apt-get install rpm
+CONFIG_FILE="/opt/aws/amazon-cloudwatch-agent/bin/cloudwatch-config.json"
+AGENT_DIR="/opt/aws/amazon-cloudwatch-agent/bin/"
 
-wget https://amazoncloudwatch-agent.s3.amazonaws.com/debian/amd64/latest/amazon-cloudwatch-agent.deb
-sudo tee /opt/aws/amazon-cloudwatch-agent/bin/cloudwatch-config.json > /dev/null <<EOF
+# Create the directory if it doesn't exist
+sudo mkdir -p "$AGENT_DIR"
+
+# Check if the directory and file exist
+if [ ! -d "$AGENT_DIR" ]; then
+    echo "Error: Directory does not exist: $AGENT_DIR"
+    exit 1
+fi
+
+# Create a JSON configuration file using sudo tee
+sudo tee "$CONFIG_FILE" > /dev/null <<EOF
 {
   "agent": {
     "metrics_collection_interval": 10,
@@ -45,9 +56,21 @@ sudo tee /opt/aws/amazon-cloudwatch-agent/bin/cloudwatch-config.json > /dev/null
   }
 }
 EOF
+
+# Verify that the JSON configuration file has been created
+if [ ! -f "$CONFIG_FILE" ]; then
+    echo "Error: Configuration file not created: $CONFIG_FILE"
+    exit 1
+fi
+wget https://amazoncloudwatch-agent.s3.amazonaws.com/debian/amd64/latest/amazon-cloudwatch-agent.deb
+
 sudo dpkg -i amazon-cloudwatch-agent.deb
 sudo /opt/aws/amazon-cloudwatch-agent/bin/amazon-cloudwatch-agent-ctl -a fetch-config -m onPremise -s -c file:/opt/aws/amazon-cloudwatch-agent/bin/cloudwatch-config.json
 sudo apt-get -f install -y
+
+sudo systemctl enable amazon-cloudwatch-agent
+sudo systemctl start amazon-cloudwatch-agent
+sudo systemctl restart amazon-cloudwatch-agent
 
 sudo unzip WebAppRenamed -d WebApp
